[#20] - remame Main.jsx to Dashboard.jsx and remove unnecessary files
